test(agent): retain single-call coverage and remove ledger recovery cases
