--- server/tests/agent/fixtures/execution_v1_outputs.sql
@@ -1,16 +1,0 @@
--- Generated through Agent.realize_action_plan at PR117 merge a866f521.
--- Four public executions: complete, completed output prefix, safe no-output failure, unknown receipt.
-CREATE TABLE agent_executions (
-	character_id VARCHAR NOT NULL,
-	execution_id VARCHAR NOT NULL,
-	version INTEGER NOT NULL,
-	fingerprint VARCHAR NOT NULL,
-	plan_json TEXT NOT NULL,
-	facts_json TEXT NOT NULL,
-	occupied INTEGER NOT NULL,
-	PRIMARY KEY (character_id, execution_id)
-);
-INSERT INTO "agent_executions" VALUES('luotianyi','legacy_terminal',1,'b981887be18f07f9d9f0cfe9b77402a28fbf1cb24a92ba81d78df257fa2a3e90','[1,["ActionPlan",{"actions":["tuple",[["Say",{"action_id":"a2","content":"你好","delivery":["OutputDelivery","conversation"],"expression":null,"prepared_audio_ref":null,"sound_content":null,"tone":["Tone",{"value":"normal"}]}],["Sing",{"action_id":"a1","expression":null,"segment_id":"verse","song_id":"song"}]]],"basis_interaction_revision":3,"interaction_id":"i","origin_request_id":"r","plan_id":"p","plan_ordinal":1,"source_stimulus_ids":["tuple",["m2","m1"]],"target_character_id":"luotianyi"}]]','[{"confirmed":true,"result":{"action_id":"a2","effect_ref":null,"error_code":null,"irreversible_effect_committed":false,"status":"completed"},"started":true,"unknown":false},{"confirmed":true,"result":{"action_id":"a1","effect_ref":null,"error_code":null,"irreversible_effect_committed":false,"status":"completed"},"started":true,"unknown":false}]',0);
-INSERT INTO "agent_executions" VALUES('luotianyi','legacy_partial',1,'b981887be18f07f9d9f0cfe9b77402a28fbf1cb24a92ba81d78df257fa2a3e90','[1,["ActionPlan",{"actions":["tuple",[["Say",{"action_id":"a2","content":"你好","delivery":["OutputDelivery","conversation"],"expression":null,"prepared_audio_ref":null,"sound_content":null,"tone":["Tone",{"value":"normal"}]}],["Sing",{"action_id":"a1","expression":null,"segment_id":"verse","song_id":"song"}]]],"basis_interaction_revision":3,"interaction_id":"i","origin_request_id":"r","plan_id":"p","plan_ordinal":1,"source_stimulus_ids":["tuple",["m2","m1"]],"target_character_id":"luotianyi"}]]','[{"confirmed":true,"result":{"action_id":"a2","effect_ref":null,"error_code":null,"irreversible_effect_committed":false,"status":"completed"},"started":true,"unknown":false},{"confirmed":false,"result":{"action_id":"a1","effect_ref":null,"error_code":"PROVIDER_TIMEOUT","irreversible_effect_committed":false,"status":"failed"},"started":true,"unknown":false}]',0);
-INSERT INTO "agent_executions" VALUES('luotianyi','legacy_no_output',1,'b981887be18f07f9d9f0cfe9b77402a28fbf1cb24a92ba81d78df257fa2a3e90','[1,["ActionPlan",{"actions":["tuple",[["Say",{"action_id":"a2","content":"你好","delivery":["OutputDelivery","conversation"],"expression":null,"prepared_audio_ref":null,"sound_content":null,"tone":["Tone",{"value":"normal"}]}],["Sing",{"action_id":"a1","expression":null,"segment_id":"verse","song_id":"song"}]]],"basis_interaction_revision":3,"interaction_id":"i","origin_request_id":"r","plan_id":"p","plan_ordinal":1,"source_stimulus_ids":["tuple",["m2","m1"]],"target_character_id":"luotianyi"}]]','[{"confirmed":false,"result":{"action_id":"a2","effect_ref":null,"error_code":"PROVIDER_TIMEOUT","irreversible_effect_committed":false,"status":"failed"},"started":true,"unknown":false},{"confirmed":false,"result":null,"started":false,"unknown":false}]',0);
-INSERT INTO "agent_executions" VALUES('luotianyi','legacy_unknown',1,'b981887be18f07f9d9f0cfe9b77402a28fbf1cb24a92ba81d78df257fa2a3e90','[1,["ActionPlan",{"actions":["tuple",[["Say",{"action_id":"a2","content":"你好","delivery":["OutputDelivery","conversation"],"expression":null,"prepared_audio_ref":null,"sound_content":null,"tone":["Tone",{"value":"normal"}]}],["Sing",{"action_id":"a1","expression":null,"segment_id":"verse","song_id":"song"}]]],"basis_interaction_revision":3,"interaction_id":"i","origin_request_id":"r","plan_id":"p","plan_ordinal":1,"source_stimulus_ids":["tuple",["m2","m1"]],"target_character_id":"luotianyi"}]]','[{"confirmed":false,"result":{"action_id":"a2","effect_ref":null,"error_code":"PROVIDER_TIMEOUT","irreversible_effect_committed":false,"status":"failed"},"started":true,"unknown":true},{"confirmed":false,"result":null,"started":false,"unknown":false}]',0);
