--- server/tests/agent/test_request_storage_recovery.py
@@ -1,102 +1,0 @@
-"""真实 SQL 存储故障注入；仅通过公开 handle 检查损坏记录不触发重跑。"""
-import json
-
-import pytest
-from sqlalchemy import MetaData, Table, event, select, update
-
-import src.domain.agent as d
-from routing_support import Sink, plan_and_context, request, settlement
-from plan_emission_support import draft
-
-
-async def _deliver(req, plans):
-    plan, _ = plan_and_context()
-    receipt = await plans.emit(draft(actions=plan.actions, sources=plan.source_stimulus_ids))
-    return settlement(req, emitted=(receipt.plan_id,))
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("damage", [
-    "row_version", "fingerprint", "invalid_json", "report_version", "missing_field",
-    "status", "identity_tuple", "wrong_request", "wrong_trigger", "wrong_revision", "wrong_pending",
-])
-async def test_corrupt_persisted_record_refuses_replay_without_reprocessing(
-    routed_runtime, runtime_dependencies, damage,
-):
-    runtime, _ = routed_runtime(handle=_deliver)
-    await runtime.get_agent().handle_stimulus(request(), Sink())
-    await runtime.shutdown()
-    kwargs, _ = runtime_dependencies
-    sessions = kwargs["database_manager"].open_sql_session
-    # SQL 是外部故障注入边界；表仅用于破坏输入，不断言账本内部查询步骤。
-    with sessions() as session:
-        table = Table("agent_handle_requests", MetaData(), autoload_with=session.get_bind())
-        key = (table.c.character_id == "luotianyi") & (table.c.request_id == "r")
-        payload = session.execute(select(table.c.report_json).where(key)).scalar_one()
-        data = json.loads(payload)
-        changes = {}
-        if damage == "row_version":
-            changes["version"] = 999
-        elif damage == "fingerprint":
-            changes["fingerprint"] = "broken"
-        elif damage == "invalid_json":
-            changes["report_json"] = "{"
-        else:
-            if damage == "report_version":
-                data["version"] = 999
-            elif damage == "missing_field":
-                del data["report"]["retryable"]
-            else:
-                name, value = {
-                    "status": ("request_status", "unknown"),
-                    "identity_tuple": ("emitted_plan_ids", "p"),
-                    "wrong_request": ("request_id", "other"),
-                    "wrong_trigger": ("trigger_stimulus_id", "other"),
-                    "wrong_revision": ("basis_interaction_revision", 999),
-                    "wrong_pending": ("considered_pending_stimulus_ids", ["other"]),
-                }[damage]
-                data["report"][name] = value
-            changes["report_json"] = json.dumps(data)
-        session.execute(update(table).where(key).values(**changes))
-        session.commit()
-    replacement, _ = routed_runtime(handle=_deliver)
-    sink = Sink()
-    for _ in range(2):
-        report = await replacement.get_agent().handle_stimulus(request(), sink)
-        assert report.request_status is d.HandlingRequestStatus.FAILED
-        assert report.error_code is d.HandlingErrorCode.DEPENDENCY_UNAVAILABLE
-        assert report.retryable is False
-    assert sink.values == []
-
-
-@pytest.mark.asyncio
-async def test_commit_failure_logs_only_the_returned_failed_settlement(routed_runtime, runtime_dependencies, caplog):
-    from src.utils.logger import get_logger
-
-    kwargs, _ = runtime_dependencies
-    sessions = kwargs["database_manager"].open_sql_session
-    reject_settlement = False
-
-    def before_commit(session):
-        if reject_settlement:
-            raise RuntimeError("commit unavailable")
-
-    async def handle(req, plans):
-        nonlocal reject_settlement
-        result = await _deliver(req, plans)
-        reject_settlement = True
-        return result
-
-    runtime, _ = routed_runtime(handle=handle)
-    logger = get_logger("src.agent.facade")
-    logger.addHandler(caplog.handler)
-    event.listen(sessions, "before_commit", before_commit)
-    try:
-        report = await runtime.get_agent().handle_stimulus(request(), Sink())
-        assert report.error_code is d.HandlingErrorCode.DEPENDENCY_UNAVAILABLE
-        settlements = [record.getMessage() for record in caplog.records if "Agent settlement" in record.getMessage()]
-        assert len(settlements) == 1
-        assert "status=failed" in settlements[0]
-    finally:
-        event.remove(sessions, "before_commit", before_commit)
-        logger.removeHandler(caplog.handler)
