+++ server/tests/conftest.py
@@ -6,16 +6,12 @@ import pytest
 
 _ACTIVE_TEST_FILES = {
     (Path(__file__).parent / "agent" / "test_output_sequences.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_output_delivery_recovery.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_output_storage_recovery.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_output_recovery_lifecycle.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_execution_idempotency.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_execution_storage_recovery.py").resolve(),
+    (Path(__file__).parent / "agent" / "test_output_delivery_failure.py").resolve(),
+    (Path(__file__).parent / "agent" / "test_execution_isolation.py").resolve(),
     (Path(__file__).parent / "agent" / "test_plan_emission.py").resolve(),
     (Path(__file__).parent / "agent" / "test_plan_logging.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_plan_delivery_recovery.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_request_idempotency.py").resolve(),
-    (Path(__file__).parent / "agent" / "test_request_storage_recovery.py").resolve(),
+    (Path(__file__).parent / "agent" / "test_plan_delivery_failure.py").resolve(),
+    (Path(__file__).parent / "agent" / "test_request_isolation.py").resolve(),
     (Path(__file__).parent / "system" / "test_system_runtime_shutdown.py").resolve(),
     (Path(__file__).parent / "agent" / "test_handler_registration.py").resolve(),
     (Path(__file__).parent / "agent" / "test_handler_dispatch.py").resolve(),
